server/auth/dao/mongo.go
package dao

import (
	"context"
	"fmt"

	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"go.mongodb.org/mongo-driver/mongo"
	"go.mongodb.org/mongo-driver/mongo/options"
)

type Mongo struct {
	col *mongo.Collection
}

func NewMongo(db *mongo.Database) *Mongo {
	return &Mongo{
		col: db.Collection("account"),
	}
}

func (m *Mongo) ResolveAccountID(c context.Context, OpenID string) (string, error) {
	res := m.col.FindOneAndUpdate(c, bson.M{
		"open_id": OpenID,
	}, bson.M{
		"$set": bson.M{
			"oepn_id": OpenID,
		},
	}, options.FindOneAndUpdate().SetUpsert(true).SetReturnDocument(options.After))

	if err := res.Err(); err != nil {
		return "", fmt.Errorf("connot FindOneAndUpdate: %v", err)
	}

	var row struct {
		ID primitive.ObjectID `bson:"_id"`
	}
	err := res.Decode(&row)

	if err != nil {
		return "", fmt.Errorf("connot decode result: %v", err)
	}

	return row.ID.Hex(), nil
}
